move functions related to create test resources to a separate file
